Move Term and Term' into Data.Term.Types.
